requirements.txt
python-telegram-bot==21.4
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5
